Expect users to provide the mypy library
Users are expected to provide mypy as this project cannot provide a version
which matches all possible user use cases.
Until now users had to provide a resolved requirements file. This is
however an unnecessarily complex model. It exposes to the user that this
projects uses rules_python to manage its dependencies in a specific way.
The more natural solution is that we expect the user to provide a mypy
library matching the users Python toolchain and expectations.
On top this eases integrating this project into other workspaces as less
interaction via macros is required.