Decouple users from internal mypy for testing
Using the internal testing mypy as default for the mypy label_flag forces
users to resolve `mypy_integration_pip_deps` which is not supposed to be
relevant for them. To prevent this we set an empty library as default for
the mypy label_flag to ensure users provide their own mypy and to
decouple them from `mypy_integration_pip_deps`.