Use modern rules_python and expect users to provide the mypy library themselves (#98)
* chore: Adapt gitignore for CLion users

* chore: Remove export pointing to non existing code

* Use existing maybe helper instead of manually implementing it

We don't use `maybe` for py_repositories.bzl as maybe can't be invoked
recursively and the functions provided by rules_python already use it
internally

* Consistent construction of URLs to dependencies

While using a variable to construct the URLs is the approach with
lower maintenance, using plain strings is required to allow Renovate to
understand them.

* Update rules_python to 0.18.1

This is a recent version supporting Bazel 6 and the forward path for
managing external dependencies based on `pip_parse`. We don't use the
latest version as rules_python >= 0.19.0 is no longer compatible to
Bazel 4.x and it might still be too early for some users of this project
for this step.

Noteworthy changes:
- Use `pip_parse` instead of deprecated `pip_install`. Consequently, users
  now have to provide a fully resolved requirements file for mypy. This means,
  there is no more default mypy shipped with this project. Users always
  have to provide their own mypy. The mypy tracked in third_party is used
  solely for testing.
- We use a hermetic Python 3.8 toolchain from rules_python. Strictly speaking
  a resolved requirements file is only valid for the Python interpreter and
  the platform it was created for. As far as we now mypy has no platform
  specific dependencies. And even if this changes, rules_python offers
  a feature for registering platform specific requirement files. However,
  some mypy versions have dependencies which are only valid for specific
  Python versions (e.g. `typed-ast`). To ensure the project behaves the same
  for each developer, we use a hermetic Python toolchain fixing the Python
  version independent from the Python interpreter on the developer host.

* Expect users to provide the mypy library

Users are expected to provide mypy as this project cannot provide a version
which matches all possible user use cases.
Until now users had to provide a resolved requirements file. This is
however an unnecessarily complex model. It exposes to the user that this
projects uses rules_python to manage its dependencies in a specific way.
The more natural solution is that we expect the user to provide a mypy
library matching the users Python toolchain and expectations.
On top this eases integrating this project into other workspaces as less
interaction via macros is required.

* chore: Remove unused file

* Decouple users from internal mypy for testing

Using the internal testing mypy as default for the mypy label_flag forces
users to resolve `mypy_integration_pip_deps` which is not supposed to be
relevant for them. To prevent this we set an empty library as default for
the mypy label_flag to ensure users provide their own mypy and to
decouple them from `mypy_integration_pip_deps`.

* chore: Update rules_python to 0.27.1

* Fix grammar in doc