Extend bcr tests
The tests in the examples directory ensure the bazelized mypy understands
a lot of Python code examples. This is great, but is missing testing for
expected failures. The test script in the root module does perform such
tests.
Ideally, there would be a directory with all relevant integration tests
which uses the root module as external dependency. For now this largish
refactoring is out of scope and thus we test both the root module itself
and the examples.