Prepare next release (#131)
* Remove compatibility_level from examples

This is superfluous noise as we don't deploy the examples

* Set module to default compatibility_level

The default value is 0. While the next release will be a breaking change,
imho the compatibility_level is counted starting with the first bzlmod
release, which is still pending.
The compatibility_level is remove completely, as the bazel-central-registry
maintainers don't want modules to explicitly mention the default value.

* Use rolling releases in bcr tests

We should be aware of incompatibilities to upcoming Bazel versions as early
as possible.
We reorder the arguments as increasing version from left to right seems to
be the dominant style in the bcr.

* Extend bcr tests

The tests in the examples directory ensure the bazelized mypy understands
a lot of Python code examples. This is great, but is missing testing for
expected failures. The test script in the root module does perform such
tests.
Ideally, there would be a directory with all relevant integration tests
which uses the root module as external dependency. For now this largish
refactoring is out of scope and thus we test both the root module itself
and the examples.

* Rework release notes template

- Advertise bzlmod as preferred option due to WORKSPACE being legacy
- Do not explain how to use bzlmod. The information is not sufficient to
  understand properly what bzlmod is about for somebody new to bzlmod.
  The information is superfluous for everybody who has read the bzlmod doc.
- Fix WORKSPACE template to work with current HEAD